ref: minor code refactoring
