+++ go.mod
@@ -1,0 +1,4 @@
+module github.com/teadove/gobrot
+
+go 1.20
+
