+++ internal/palette/color.go
