+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/teadove/gobrot/internal/cli"
+
+func main() {
+	cli.Run()
+}
